+++ tests/Integration/CachedBuilder/LazyLoadTest.php
@@ -3,7 +3,6 @@
 use GeneaLabs\LaravelModelCaching\Tests\Fixtures\Author;
 use GeneaLabs\LaravelModelCaching\Tests\Fixtures\Book;
 use GeneaLabs\LaravelModelCaching\Tests\Fixtures\Profile;
-use GeneaLabs\LaravelModelCaching\Tests\Fixtures\Store;
 use GeneaLabs\LaravelModelCaching\Tests\Fixtures\UncachedAuthor;
 use GeneaLabs\LaravelModelCaching\Tests\Fixtures\UncachedBook;
 use GeneaLabs\LaravelModelCaching\Tests\Fixtures\UncachedProfile;
@@ -39,37 +38,6 @@ class LazyLoadTest extends IntegrationTestCase
         $this->assertNotNull($result);
         $this->assertNotNull($cachedResult);
         $this->assertNotNull($uncachedResult);
-    }
-
-    public function testBelongsToManyRelationship()
-    {
-        $bookId = (new Store)
-            ->disableModelCaching()
-            ->with("books")
-            ->first()
-            ->books
-            ->first()
-            ->id;
-        $key = sha1("genealabs:laravel-model-caching:testing::memory::book-store:genealabslaravelmodelcachingcachedbelongstomany-book_store.book_id_=_{$bookId}");
-        $tags = [
-            'genealabs:laravel-model-caching:testing::memory::genealabslaravelmodelcachingtestsfixturesstore',
-        ];
-
-        $stores = (new Book)
-            ->find($bookId)
-            ->stores;
-        $cachedStores = $this
-            ->cache()
-            ->tags($tags)
-            ->get($key)['value'];
-        $uncachedBook = (new UncachedBook)
-            ->find($bookId);
-        $uncachedStores = $uncachedBook->stores;
-
-        $this->assertEquals($uncachedStores->pluck("id"), $stores->pluck("id"));
-        $this->assertEquals($uncachedStores->pluck("id"), $cachedStores->pluck("id"));
-        $this->assertNotNull($cachedStores);
-        $this->assertNotNull($uncachedStores);
     }
 
     public function testHasManyRelationship()